+++ rust-engine/src/phase/explicit_monadic.rs
@@ -65,7 +65,7 @@ impl ExplicitMonadicVisitor {
                     output: expr.ty.clone(),
                 })),
                 meta: Metadata {
-                    span: expr.meta.span.clone(),
+                    span: expr.meta.span,
                     attributes: vec![],
                 },
             },
@@ -181,8 +181,7 @@ impl ExplicitMonadicVisitor {
             }
             ExprKind::Assign { value: inner, .. }
             | ExprKind::Borrow { inner, .. }
-            | ExprKind::AddressOf { inner, .. }
-            | ExprKind::Deref(inner) => {
+            | ExprKind::AddressOf { inner, .. } => {
                 self.visit_expr_coerce(MonadicStatus::Value, inner);
                 Some(MonadicStatus::Value)
             }