+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Changes to the Rust Engine:
  - Remove `BinOp` resugaring (#1950)
  - Apply resugarings to linked items (pre/post conditions) (#1961)
+ - Add new import_thir implemented in Rust and using `FullDef`, activated with `--experimental-full-def` (#1967)
 
 Changes to the frontend: